personal-chat-bot/chatbot-personal/backend/requirements.txt
django
djangorestframework
psycopg2-binary
pgvector
openai
python-dotenv

